streaming/requirements.txt
kafka-python
faker
pandas
schedule
